filter slice

+++ src/components/ContactList/ContactList.jsx
@@ -5,12 +5,18 @@ import { deleteContact } from 'redux/slice/contactsSlice';
 
 export const ContactList = () => {
   const allContacts = useSelector(state => state.contacts);
+  const filterName = useSelector(state => state.filter);
 
   const dispatch = useDispatch();
 
+  const normalizedFilter = filterName.toLowerCase();
+  const filteredContacts = allContacts.filter(contact =>
+    contact.name.toLowerCase().includes(normalizedFilter)
+  );
+
   return (
     <ul>
-      {allContacts.map(contact => {
+      {filteredContacts.map(contact => {
         const nameArray = contact.name.split(' ');
         const nameFirstUpperLetter = nameArray
           .map(word => word.replace(word[0], word[0].toUpperCase()))
@@ -24,7 +30,11 @@ export const ContactList = () => {
               type="ButtonDelete"
               id={contact.id}
               onClick={() => {
-                dispatch(deleteContact(contact.id));
+                if (
+                  window.confirm('Are you sure you want to delete the contact?')
+                ) {
+                  dispatch(deleteContact(contact.id));
+                }
               }}
             >
               Delete
